Data on state aid awards

This dataset includes data on state aid awards granted to firms by government agencies in member states that were reported to the Commission. There is one observation per award (2016-2020). The dataset includes information on the date of the award, the aid-granting agency, the beneficiary, the NACE sector of the beneficiary, the aid instrument used, and the estimated amount of the aid in euros.

awards

Format

A data frame with 34 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

case_number

String. A number that uniquely identifies each state aid case. Assigned by the Commission. The Commission changed the format of case numbers in 2010. Before this change, the case number indicated the type(s) of the procedure associated with the case. After this change, all case numbers have the format SA.#####.

reference_number

String. An ID number that uniquely identifies each state aid award.

notification_date

Date. The date the member state notified the Commission of the state aid measure in the format YYYY-MM-DD.

notification_year

Numeric. The year the member state notified the Commission of the state aid measure.

notification_month

Numeric. The month the member state notified the Commission of the state aid measure.

notification_day

Numeric. The day the member state notified the Commission of the state aid measure.

publication_date

Date. The date that the Commission published a record of the state aid award in the format YYYY-MM-DD.

publication_year

Numeric. The year that the Commission published a record of the state aid award.

publication_month

Numeric. The month that the Commission published a record of the state aid award.

publication_day

Numeric. The day that the Commission published a record of the state aid award.

member_state_id

Numeric. An ID number that uniquely identifies each member state. This ID number is assigned when member states are sorted by accession date and then alphabetically.

member_state

String. The name of the member state that the Commission opened the case against.

member_state_code

String. A two letter code assigned by the Commission that uniquely identifies each member state.

authority_name

String. The name of the national, subnational, or local authority that granted the state aid award. Note that the text of this variable has not been cleaned.

region

String. The region of the member state where the beneficiary of the state aid award is located. Note that the text of this variable has not been cleaned.

beneficiary_name

String. The name of the beneficiary of the state aid award. Note that the text of this variable has not been cleaned.

beneficiary_type_id

Numeric. An ID number that uniquely identifies each type of beneficiary. Coded 1 for Small or medium-sized enterprise (SME) and 2 for Large enterprise.

beneficiary_type

String. The type of the beneficiary of the state aid award. Coded Small or medium-sized enterprise (SME) or Large enterprise. Uses the Commission’s standard definition of small and median-sized enterprises (SMEs).

nace_sector_id

Numeric. An ID number that uniquely identifies each NACE sector.

nace_sector

String. The NACE sector for the state aid award.

nace_sector_code

String. The NACE sector code for the state aid award (a single capital letter).

nace_code

String. The NACE code for the state aid award.

nace_description

String. A description of the NACE code for the state aid award.

aid_instrument_id

Numeric. An ID number that uniquely identifies each state aid instrument that a granting authority can use.

aid_instrument

String. The aid instrument that the granting authority used in granting the state aid award.

raw_amount

Numeric. The amount of the state aid award expressed in units of the local currency. If the value of the award is expressed as a range, this variable records the mean of the range.

currency

String. The name of the currency in which the member state reported the amount of the state aid award.

range

Dummy. A dummy variable indicating whether the amount of the state aid award is expressed as a range instead of a specific value (allowed by the Commission under some circumstances to protect proprietary business data). Coded 1 if the amount is expressed as a range and 0 otherwise.

range_min

Numeric. If the value of the state aid award is expressed as a range (see range), the minimum value of the range expressed in units of the local currency. Coded NA if the value of the award is not expressed as a range.

range_max

Numeric. If the value of the state aid award is expressed as a range (see range), the maximum value of the range expressed in units of the local currency. Coded NA if the value of the award is not expressed as a range.

exchange_rate

Numeric. The exchange rate for converting the local currency into euros. Coded 1 if the local currency is the euro.

amount_euros

Numeric. The total estimated value of the award in euros.

voluntary

Dummy. A dummy variable indicating whether the state aid award was small enough that notification to the Commission is voluntary. Coded 1 if notification is voluntary and 0 otherwise.