Data on cases

This dataset include data on state aid cases. There is one observation per case (1988-2020). The dataset includes information on the department responsible for the case, the member state that the case is against, and decisions taken by the Commission in the case.

cases

Format

A data frame with 42 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

case_number

String. A number that uniquely identifies each state aid case. Assigned by the Commission. The Commission changed the format of case numbers in 2010. Before this change, the case number indicated the type(s) of the procedure associated with the case. After this change, all case numbers have the format SA.#####.

procedure_numbers

String. Numbers that uniquely identifiy each procedure associated with a case. If there are multiple procedures, the procedure numbers are listed and separated by a comma. After the Commission changed the format of case numbers in 2010, only procedure numbers indicate the types of procedures associated with a case.

member_state_id

Numeric. An ID number that uniquely identifies each member state. This ID number is assigned when member states are sorted by accession date and then alphabetically.

member_state

String. The name of the member state that the Commission opened the case against.

member_state_code

String. A two letter code assigned by the Commission that uniquely identifies each member state.

department_id

Numeric. An ID number that uniquely identifies each Directorate-General (DG) of the Commission. There are three DGs that can open state aid cases. The DG for Competition (COMP) is coded 1, the DG for Agriculture and Rural Development (AGRI) is coded 2, and the DG for Maritime Affairs and Fisheries (MARE) is coded 3.

department

String. The name of the Directorate-General (DG) of the Commission that opened the state aid case.

department_code

String. A multi-letter code assigned by the Commission that uniquely identifies each Directorate-General (DG) of the Commission.

case_type_id

Numeric. An ID number that uniquely identifies each type of state aid cases. Coded 1 for cases that involve state aid measures granted through Commission-approved schemes that is not notifiable, coded 2 for cases that involve state aid measures granted through Commission-approved schemes where the Commission has required that the member state notify any aid granted through the scheme, and coded 3 for cases that involve ad hoc state aid measures that are not granted through a scheme.

case_type

String. The type of the state aid case. There are three types of cases. Coded Scheme for cases that involve state aid measures granted through Commission-approved schemes that is not notifiable, coded Individual application for cases that involve state aid measures granted through Commission-approved schemes where the Commission has required that the member state notify any aid granted through the scheme, and coded Ad hoc for cases that involve ad hoc state aid measures that are not granted through a scheme.

procedure_types

String. A list of the types of procedures associated with the case, separated by a comma. Procedure types are listed in the order that the Commission opened the procedures. Possible values include: Contradictory aid, for state aid that the Commission suspects violates the rules of the internal market, as defined by Article 107 of the Treaty of the Functioning of the European Union (TFEU); Existing aid, for existing state aid measures; Notified aid, for state aid that the member state has notified to the Commission; Unnotified aid, for state aid that the member state has not notified to the Commission; General block exemption, for state aid that is exempt from notification under the General Block Exemption Regulation (GBER); and Special block exemption, for state aid that is exempt from notification under a specific block exemption.

contradictory_aid

Dummy. A dummy variable indicating whether the case relates to whether the Commission suspects that the state aid measure is contradictory to the rules of the single market, as defined by Article 107 of the Treaty of the Functioning of the European Union (TFEU). Coded 1 if it does and 0 if it does not.

existing_aid

Dummy. A dummy variable indicating whether the case relates to existing state aid measures. Coded 1 if it does and 0 if it does not.

notified_aid

Dummy. A dummy variable indicating whether the case relates to state aid notified by the member state. Coded 1 if it does and 0 if it does not.

unnotified_aid

Dummy. A dummy variable indicating whether the case relates to state aid that is not notified by the member state. Coded 1 if it does and 0 if it does not.

general_block_exemption

Dummy. A dummy variable indicating whether the case relates to state that is exempt from notification under the General Block Exemption Regulation (GBER). Coded 1 if it does and 0 if it does not.

specific_block_exemption

Dummy. A dummy variable indicating whether the case relates to state aid that is exempt from notification under a specific block exemption. Coded 1 if it does and 0 if it does not.

notification_date

Date. The date the member state notified the Commission of the state aid measure in the format YYYY-MM-DD.

notification_year

Numeric. The year the member state notified the Commission of the state aid measure.

notification_month

Numeric. The month the member state notified the Commission of the state aid measure.

notification_day

Numeric. The day the member state notified the Commission of the state aid measure.

outcome_date

Date. The date the Commission took its last decision in the case in the format YYYY-MM-DD.

outcome_year

Numeric. The year the Commission took its last decision in the case.

outcome_month

Numeric. The month the Commission took its last decision in the case.

outcome_day

Numeric. The day the Commission took its last decision in the case.

decisions

String. A list of all of the decisions made by the Commission in the case, separated by a semicolon.

count_decisions

Numeric. The total number of decisions made by the Commission in the case.

outcome

String. The outcome of the case. Possible values include: Exempt from notification, Does not constitute aid, No objectionPositive Decision, Negative Decision, Conditional decision. Coded Notification withdrawn if the member state withdrew the notification, either before the formal investigation procedure under Article 10(1) of Council Regulation (EU) 2015/1589, formerly Article 8(1) of Council Regulation (EC) No 659/1999, or during the formal investigation procedure under Article 10(2) of Council Regulation (EU) 2015/1589, formerly Article 8(2) of Council Regulation (EC) No 659/1999.

outcome_phase_1

String. If the Commission opened a preliminary investigation, the outcome of that phase of the procedure. Possible values include: Does not constitute aid, No objection, Formal investigation, and Notification withdrawn. Coded Not applicable if the Commission did not open a preliminary investigation, which is when it determines that the state aid measure is exempt from notification under Article 108(2) of the Treaty on the Functioning of the European Union (TFEU).

outcome_phase_2

String. If the Commission opened a formal investigation, the outcome of that phase of the procedure. Possible values include: Conditional decision, Does not constitute aid, Negative decision, Notification withdrawn, and Positive decision. Coded Not applicable if the Commission did not open a formal investigation under Article 4(4) of Council Regulation (EU) 2015/1589, formerly Article 4(4) of Council Regulation (EC) No 659/1999. Coded Missing record for a small number of observations where the Commission does not report data.

exempt

Dummy. A dummy variable indicating whether the state aid measure is exempt from notification under Article 108(2) of the Treaty on the Functioning of the European Union (TFEU).

preliminary_investigation

Dummy. A dummy variable indicating whether the Commission decided to open a preliminary investigation under Article 108 of the Treaty on the Functioning of the European Union (TFEU).

formal_investigation

Dummy. A dummy variable indicating whether the Commission decided to open a formal investigation under Article 4(4) of Council Regulation (EU) 2015/1589, formerly Article 4(4) of Council Regulation (EC) No 659/1999.

no_objection

Dummy. A dummy variable indicating whether the Commission decided not to raise an objection to the state aid measure under Article 4(3) of Council Regulation (EU) 2015/1589, formerly Article 4(3) of Council Regulation (EC) No 659/1999.

not_aid

Dummy. A dummy variable indicating whether the Commission decided that the state aid measure did not constitute state aid, according to Article 107 of the Treaty on the Functioning of the European Union (TFEU), under Article 4(2) of Council Regulation (EU) 2015/1589, formerly Article 4(2) of Council Regulation (EC) No 659/1999.

positive

Dummy. A dummy variable indicating whether, after a formal investigation, the Commission decided under article 9(3) of Council Regulation (EU) 2015/1589, formerly Article 7(3) of Council Regulation (EC) No 659/1999, to approve the state aid measure because it was compatible with the rules of the single market, as defined by Article 107 of the Treaty of the Functioning of the European Union (TFEU).

negative

Dummy. A dummy variable indicating whether, after a formal investigation, the Commission decided under article 9(5) of Council Regulation (EU) 2015/1589, formerly Article 7(5) of Council Regulation (EC) No 659/1999, to not approve the state aid measure because it was not compatible with the rules of the single market, as defined by Article 107 of the Treaty of the Functioning of the European Union (TFEU).

conditional

Dummy. A dummy variable indicating whether, after a formal investigation, the Commission decided under article 9(4) of Council Regulation (EU) 2015/1589, formerly Article 7(4) of Council Regulation (EC) No 659/1999, to approve the state aid measure subject to conditions to make the measure compatible with the rules of the single market, as defined by Article 107 of the Treaty of the Functioning of the European Union (TFEU).

withdrawal

Dummy. A dummy variable indicating whether the member state that notified the state aid measure withdrew the measure, either before the formal investigation procedure under Article 10(1) of Council Regulation (EU) 2015/1589, formerly Article 8(1) of Council Regulation (EC) No 659/1999, or during the formal investigation procedure under Article 10(2) of Council Regulation (EU) 2015/1589, formerly Article 8(2) of Council Regulation (EC) No 659/1999.

referral

Dummy. A dummy variable indicating whether, after a formal investigation, the Commission decided to refer the state aid case to the Court of Justice of the European Union (CJEU) under Article 14 of Council Regulation (EU) 2015/1589, formerly Article 12 of Council Regulation (EC) No 659/1999, for noncompliance with an injunction, under Article 108(2) of the Treaty on the Functioning of the European Union (TFEU), for noncompliance with a Commission decision, or under Article 260(2) of the Treaty on the Functioning of the European Union (TFEU), for noncompliance with a judgment of the Court.

recovery

Dummy. A dummy variable indicating whether, after a formal investigation, the Commission decided under 16(1) of Council Regulation (EU) 2015/1589, formerly 14(1) of Council Regulation (EC) No 659/1999, that the member state that notified that state aid measure must take all necessary measures to recover aid from the beneficiary.