CSTS data on cases (by member state and by case type)

This dataset includes aggregated data on the number of state aid cases per member state per year (cross-sectional time-series data) broken down by case type. There is one observation per member state per year per case type (1988-2020), excluding state-years where the state was not a member of the EU.

cases_csts_ms_ct

Format

A data frame with 8 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

year

Numeric. The year the case was opened by the Commission.

member_state_id

Numeric. An ID number that uniquely identifies each member state. This ID number is assigned when member states are sorted by accession date and then alphabetically.

member_state

String. The name of the member state that the Commission opened the case against.

member_state_code

String. A two letter code assigned by the Commission that uniquely identifies each member state.

case_type_id

Numeric. An ID number that uniquely identifies each type of state aid cases. Coded 1 for cases that involve state aid measures granted through Commission-approved schemes that is not notifiable, coded 2 for cases that involve state aid measures granted through Commission-approved schemes where the Commission has required that the member state notify any aid granted through the scheme, and coded 3 for cases that involve ad hoc state aid measures that are not granted through a scheme.

case_type

String. The type of the state aid case. There are three types of cases. Coded Scheme for cases that involve state aid measures granted through Commission-approved schemes that is not notifiable, coded Individual application for cases that involve state aid measures granted through Commission-approved schemes where the Commission has required that the member state notify any aid granted through the scheme, and coded Ad hoc for cases that involve ad hoc state aid measures that are not granted through a scheme.

count_cases

Numeric. A count of the number of cases opened by the Commission at this level of aggregation.