Data on decisions

This dataset include data on decisions in state aid cases. There is one observation per decision per case (1988-2020). The dataset includes information on the department responsible for the decision and the member state that the decision is against.

decisions

Format

A data frame with 30 variables:

key_id

Numeric. An ID number that uniquely identifies each observation in the dataset.

case_number

String. A number that uniquely identifies each state aid case. Assigned by the Commission. The Commission changed the format of case numbers in 2010. Before this change, the case number indicated the type(s) of the procedure associated with the case. After this change, all case numbers have the format SA.#####.

procedure_numbers

String. Numbers that uniquely identifiy each procedure associated with a case. If there are multiple procedures, the procedure numbers are listed and separated by a comma. After the Commission changed the format of case numbers in 2010, only procedure numbers indicate the types of procedures associated with a case.

member_state_id

Numeric. An ID number that uniquely identifies each member state. This ID number is assigned when member states are sorted by accession date and then alphabetically.

member_state

String. The name of the member state that the Commission opened the case against.

member_state_code

String. A two letter code assigned by the Commission that uniquely identifies each member state.

department_id

Numeric. An ID number that uniquely identifies each Directorate-General (DG) of the Commission. There are three DGs that can open state aid cases. The DG for Competition (COMP) is coded 1, the DG for Agriculture and Rural Development (AGRI) is coded 2, and the DG for Maritime Affairs and Fisheries (MARE) is coded 3.

department

String. The name of the Directorate-General (DG) of the Commission that opened the state aid case.

department_code

String. A multi-letter code assigned by the Commission that uniquely identifies each Directorate-General (DG) of the Commission.

case_type_id

Numeric. An ID number that uniquely identifies each type of state aid cases. Coded 1 for cases that involve state aid measures granted through Commission-approved schemes that is not notifiable, coded 2 for cases that involve state aid measures granted through Commission-approved schemes where the Commission has required that the member state notify any aid granted through the scheme, and coded 3 for cases that involve ad hoc state aid measures that are not granted through a scheme.

case_type

String. The type of the state aid case. There are three types of cases. Coded Scheme for cases that involve state aid measures granted through Commission-approved schemes that is not notifiable, coded Individual application for cases that involve state aid measures granted through Commission-approved schemes where the Commission has required that the member state notify any aid granted through the scheme, and coded Ad hoc for cases that involve ad hoc state aid measures that are not granted through a scheme.

procedure_types

String. Numbers that uniquely identified each procedure associated with a case. If there are multiple procedures, the procedure numbers are listed and separated by a comma. After the Commission changed the format of case numbers in 2010, only procedure numbers indicate the types of procedures associated with a case.

contradictory_aid

Dummy. A dummy variable indicating whether the case relates to whether the Commission suspects that the state aid measure is contradictory to the rules of the single market, as defined by Article 107 of the Treaty of the Functioning of the European Union (TFEU). Coded 1 if it does and 0 if it does not.

existing_aid

Dummy. A dummy variable indicating whether the case relates to existing state aid measures. Coded 1 if it does and 0 if it does not.

notified_aid

Dummy. A dummy variable indicating whether the case relates to state aid notified by the member state. Coded 1 if it does and 0 if it does not.

unnotified_aid

Dummy. A dummy variable indicating whether the case relates to state aid that is not notified by the member state. Coded 1 if it does and 0 if it does not.

general_block_exemption

Dummy. A dummy variable indicating whether the case relates to state that is exempt from notification under the General Block Exemption Regulation (GBER). Coded 1 if it does and 0 if it does not.

specific_block_exemption

Dummy. A dummy variable indicating whether the case relates to state aid that is exempt from notification under a specific block exemption. Coded 1 if it does and 0 if it does not.

notification_date

Date. The date the member state notified the Commission of the state aid measure in the format YYYY-MM-DD.

notification_year

Numeric. The year the member state notified the Commission of the state aid measure.

notification_month

Numeric. The month the member state notified the Commission of the state aid measure.

notification_day

Numeric. The day the member state notified the Commission of the state aid measure.

outcome_date

Date. The date the Commission took its last decision in the case in the format YYYY-MM-DD.

outcome_year

Numeric. The year the Commission took its last decision in the case.

outcome_month

Numeric. The month the Commission took its last decision in the case.

outcome_day

Numeric. The day the Commission took its last decision in the case.

decision_number

Numeric. The number of the decision within each case.

decision_type_id

Numeric. An ID number that uniquely identifies each type of decision the Commission can make in a state aid case.

decision_type

String. The type of the decision.

phase

String. The phase of the procedure. Coded Exempt for state aid that is exempt from notification, Phase 1 for decisions made during a preliminary investigation, Phase 2 for decisions made during a formal investigation, or Other for decisions that the Commission can make during either phase.